+++ rocketchat/api.py
@@ -1,10 +1,10 @@
 from rocketchat.calls.chat.send_message import SendMessage
 from rocketchat.calls.channels.get_public_rooms import GetPublicRooms
-from rocketchat.calls.channels.get_private_rooms import GetPrivateRooms
+from rocketchat.calls.groups.get_private_rooms import GetPrivateRooms
 from rocketchat.calls.channels.get_room_info import GetRoomInfo
-from rocketchat.calls.channels.get_private_room_info import GetPrivateRoomInfo
+from rocketchat.calls.groups.get_private_room_info import GetPrivateRoomInfo
 from rocketchat.calls.channels.get_history import GetRoomHistory
-from rocketchat.calls.channels.get_private_room_history import GetPrivateRoomHistory
+from rocketchat.calls.groups.get_private_room_history import GetPrivateRoomHistory
 from rocketchat.calls.auth.get_me import GetMe
