+++ rocketchat/calls/groups/get_private_room_info.py
